fix(ci): line-limit guardrail — split oversized hosted delegation and worker test files
